Modularize base.py, fixes #64
